feat: restructure application entry point and routes for improved organization and clarity
