+++ src/main.tsx
@@ -1,5 +1,5 @@
 import { createRoot } from "react-dom/client";
 import "./index.css";
-import App from "./app";
+import App from "./entry";
 
 createRoot(document.getElementById("root")!).render(<App />);
