+++ src/routes/auth/register.tsx
